move end_date tests to test_time_window_partitions
